requirements-dev.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.30.0
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
filelock==3.16.1
GeoAlchemy2==0.15.2
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
identify==2.6.1
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy==1.13
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.2
Pygments==2.18.0
pyogrio==0.10.0
pyproj==3.7.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.4
typer==0.12.5
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.6
watchfiles==0.24.0
websockets==13.0.1
